refactoring to add other types of dbs and going with an adapter pattern
