+++ src/revolve/nodes/run_tests.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from revolve.nodes import test_node
 from revolve.functions import run_pytest, get_file_list, log
 from revolve.prompts import get_run_test_prompt
 from revolve.llm import invoke_llm
